+++ src/test/java/test_data/Example.java
@@ -1,4 +1,4 @@
-package test;
+package test_data;
 
 import java.io.File;
